ported 404 (now called not-found) page and blog page
