+++ src/app/api/alphavantage/company/[ticker].ts
