+++ src/app/api/alphavantage/quote/[ticker].ts
